Add two new gdbarch methods to handle pointer <-> integer casts.
gdbarch_pointer_to_integer_cast() is used when casting a pointer to an
integer.  gdbarch_integer_to_pointer_cast() is used for the opposite
direction.  Both versions return a pointer to a 'struct value' and
take the same arguments as value_cast() with the addition of the
gdbarch.  Both methods can return NULL to revert to the default
behavior for these operations.

Add versions of these for CHERI that handle casts to and from
capabilities.

When casting between a capability and an integer of the same size
(uintcap_t and intcap_t), copy all of the bytes from the old value to
the new one and just change the type.

When casting from a smaller integer to a capability pointer, generate
a null capability with the integer value in the cursor.

When casting from a capability pointer to a smaller integer, extract
the address (cursor) and return its value.  This does mean that casts
in the debugger will always return the address instead of the offset.
This is probably fine since the compiler now does the same by default.