Fix local RISC-V assembler breakage
We lost some instructions and got the wrong encodings for others in
17ae6c909e. We also started falling afoul of shifting signed integer
literals left enough to reach the sign bit, which caused the assembler
to die with "bad RISC-V opcode (mask error)".

The definitions of instructions with the same name need to be adjacent,
but this was not true for csetbounds, so the assembler would then die
with "can't hash `csetbounds': exists". Fix the cincoffsetimm alias
whilst here, and add the missing csetboundsimm one. The same also
applies to some of the capability loads/stores.

Also fix the operands of ccall, it takes two sources and no
destinations, and drop the comment that doesn't apply given the
specification no longer has selectors.

Finally, the implicit atomics were not correctly including the AQ/RL
bits, nor did they have the AQ/RL variants defined.

This all gets the assembler in a state where it no longer dies with an
internal error.